GameLobby Partially Implemented.
This commit adds:
(a) A working master server.
(b) The ability for the host to create a game lobby and register the
game to the server.
(c) The ability for a client to query the master server about the
waiting lobbies and join one of the first five available.

What’s still missing:
Host should be prompted for a game name before hosting, for now all
games are called “Generic Game”.
Client should be able to scroll through a list of more games than just
the first five.
Both client and server should be able to back out of the lobby.
Lobby needs UI functionality implemented.